configure doomemacs
we use tangle to autogenerate config.el and packages.el from config.org,
so that's why they're excluded from version control. actually,
packages.el is not needed anymore because the plan is to install it with
nix. init.el has the starting configuration for doom, which is pretty
beefy right now. also elfeed.org has the entire list of all my RSS
feeds, exported from Inoreader. currently not working but we'll get it there.