fix infinite recursion in nixosConfigurations
i'm not sure why this was needed, but deleting the specific system files
and putting them into flake.nix solved this issue. if unique
configuration to the specific system becomes bloated (ideally not), we
can consider refactoring, but the external modules would need to be
provided in flake.nix still I guess.

this was also done for darwinConfigurations but they are untested