fix system mismatch with k9s
builtins.fetchurl is run during evaluation time, not actual build time,
so it uses the system that the nix command is being run on, not where
the target is. this made me exclude k9s before from every device because
running on aarch64-darwin gave problems when deploying to x86_64-linux
since obviously the systems don't match. pkgs.fetchurl solves this issue
but of course it has a slightly different API which is simpler because
it is a derivation