convert emacs to emacs-unstable-pgtk through emacs-overlay
there are some commented commands from the old emacs-plus@29 specific
version from homebrew. I might want to try out a macport but it doesn't
exist in emacs-overlay yet, so I guess I'll just have to wait

How do I run this as a service so it always opens on startup and I don't
have to open a terminal?