fix: filter nostatoo games by suffixed name
it seems like steam will periodically erase values that don't exist in
its spec for shortcuts, which means injecting Nostatoo=true is not a
realiable matching mechanism. This caused plenty of duplicates when the
device was restarted. appname will remain consistent however, even if it
is a little ugly. maybe it will eventually make sense to use something
else like steam-rom-manager or match on the StartDir path, which should
probably be set for each anyways