fix: replace nostatoo with steam-rom-manager
Nostatoo is helpful at being a simple UI for ad-hoc changes and queries,
but it is limited in its capabilities, and I don't want to recreate the
wheel that steam-rom-manager has already succeeded reasonably well at.

The Nix API is a little outdated here, but it's still somewhat clear
what things mean. It can be refactored later for clarity.