fix: hy3 and hyprspace plugins
the issue was breaking changes in hyprland which have since been fixed
in open PRs. allowing these changes required pinning every hyprland
dependency to master, which seems to be working just fine. we'll move
these to version branches later as soon as this round of changes goes
through in upstream