fix: install CRDs that aren't managed by charts
fromCRD only creates modules for schematized custom resources. it
doesn't actually add the CRDs to the generated manifests. some charts
manage their own CRDs but we need to import the ones that don't. the
mechanism to do this is with applications.*.yamls which are strings, so
they have to be read from the src simultaneously with option generation